pyproject.toml
[project]
name = "foundation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.57.1",
    "gradio>=5.35.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "openai>=1.93.0",
    "openai-agents>=0.1.0",
    "pypdf>=5.7.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
